Added functionality to Edit Profile Page and fixed Profile Page errors
Made the color of the sign out and save button the same. Made the real user data show up in the edit profile page and made the save button change the database. Fixed functionality on week and month filters. Changed the profile name in the top left of the profile page to update when the database does.